Record: ignore the order of the locations provided in the regions blocks
The list of locations provided in the meta of the regions block were
depending on a specific order that was difficult to reproduce, leading
to having the terraform provider finding differences where there was
not any.

This fixes it by comparing the contents of the 'georegion', 'country',
'us_state' and 'ca_province' meta by first splitting the values on
commas, and verifying that they both contain the same elements.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@shopify.com>