INT-62: Allow secondary zones to become primary in place
Old behavior was to force a new resource (DELETE/PUT) on any change
to the "primary" key (secondary zone).
New behavior is to only force a new resource when the "primary" key
is set (zone becomes secondary). That is, edits to the secondary's
primary IP, making a secondary zone "normal", or going directly from
secondary to primary are supported in place.

* customdiff function checks what we're doing to the primary key and
  sets ForceNew appropriately.
* some logic to deal with API quirks around clearing secondary when we
  are becoming primary and vice-versa.
* acc tests try to exercise all of the paths a zone might take
* Note in resource_zone docs about zone primary<->secondary gotchas.