resource_record: add support for 'allow_override'
When creating a new zone through terraform, it was not possible
to support the automatic modification of the nameservers at the
same time, as the NS records were automatically created while
creating the zone, hence making mandatory a manual intervention
to be able to edit those records afterward.

This introduces a 'allow_overwrite' boolean that, when set to
true, will try to remove a record that already exists and recreate
it right away, hence allowing to set nameservers properly.