ns1_zone: add a new 'autogenerate_ns_record' parameter defaulting to 'true'
This new parameters is only used on creating a zone and allows to
prevent for NS records to be created automatically, which would
prevent users to create those records by themselves. This does it
by actually removing the auto-created NS record when the option
is set to 'false'. By default, the option is set to 'true'. We
also ignore any change to that value after creation, as it will
not imply any change to the deployed resource.