replace examples in README with links to docs
Some of the README examples are out-of-date/wrong. We didn't really
notice because the "real" docs are part of the repo and are published
on release, and we're not usuallty looking at the the README for a
source of truth.

Rather than cleaning up the README examples and having to keep them up
to date with future changes, just replace the examples with links to
the better maintained docs.